move handle_hooks tests to appropriate file, update import in tests
