+++ stacker/tests/test_util.py
@@ -8,18 +8,16 @@ import unittest
 
 import string
 import os
-import queue
 
 import mock
 
 import boto3
 
-from stacker.config import Hook, GitPackageSource
+from stacker.config import GitPackageSource
 from stacker.util import (
     cf_safe_name,
     load_object_from_string,
     camel_to_snake,
-    handle_hooks,
     merge_map,
     yaml_to_ordered_dict,
     get_client_region,
@@ -33,10 +31,6 @@ from stacker.util import (
     SourceProcessor
 )
 
-from .factories import (
-    mock_context,
-    mock_provider,
-)
 
 regions = ["us-east-1", "cn-north-1", "ap-northeast-1", "eu-west-1",
            "ap-southeast-1", "ap-southeast-2", "us-west-2", "us-gov-west-1",
@@ -272,148 +266,6 @@ Outputs:
                 sp.determine_git_ref({'uri': 'git@foo', 'tag': 'v1.0.0'}),
                 'v1.0.0'
             )
-
-
-hook_queue = queue.Queue()
-
-
-def mock_hook(*args, **kwargs):
-    hook_queue.put(kwargs)
-    return True
-
-
-def fail_hook(*args, **kwargs):
-    return None
-
-
-def exception_hook(*args, **kwargs):
-    raise Exception
-
-
-def context_hook(*args, **kwargs):
-    return "context" in kwargs
-
-
-def result_hook(*args, **kwargs):
-    return {"foo": "bar"}
-
-
-class TestHooks(unittest.TestCase):
-
-    def setUp(self):
-        self.context = mock_context(namespace="namespace")
-        self.provider = mock_provider(region="us-east-1")
-
-    def test_empty_hook_stage(self):
-        hooks = []
-        handle_hooks("fake", hooks, self.provider, self.context)
-        self.assertTrue(hook_queue.empty())
-
-    def test_missing_required_hook(self):
-        hooks = [Hook({"path": "not.a.real.path", "required": True})]
-        with self.assertRaises(ImportError):
-            handle_hooks("missing", hooks, self.provider, self.context)
-
-    def test_missing_required_hook_method(self):
-        hooks = [{"path": "stacker.hooks.blah", "required": True}]
-        with self.assertRaises(AttributeError):
-            handle_hooks("missing", hooks, self.provider, self.context)
-
-    def test_missing_non_required_hook_method(self):
-        hooks = [Hook({"path": "stacker.hooks.blah", "required": False})]
-        handle_hooks("missing", hooks, self.provider, self.context)
-        self.assertTrue(hook_queue.empty())
-
-    def test_default_required_hook(self):
-        hooks = [Hook({"path": "stacker.hooks.blah"})]
-        with self.assertRaises(AttributeError):
-            handle_hooks("missing", hooks, self.provider, self.context)
-
-    def test_valid_hook(self):
-        hooks = [
-            Hook({"path": "stacker.tests.test_util.mock_hook",
-                  "required": True})]
-        handle_hooks("missing", hooks, self.provider, self.context)
-        good = hook_queue.get_nowait()
-        self.assertEqual(good["provider"].region, "us-east-1")
-        with self.assertRaises(queue.Empty):
-            hook_queue.get_nowait()
-
-    def test_valid_enabled_hook(self):
-        hooks = [
-            Hook({"path": "stacker.tests.test_util.mock_hook",
-                  "required": True, "enabled": True})]
-        handle_hooks("missing", hooks, self.provider, self.context)
-        good = hook_queue.get_nowait()
-        self.assertEqual(good["provider"].region, "us-east-1")
-        with self.assertRaises(queue.Empty):
-            hook_queue.get_nowait()
-
-    def test_valid_enabled_false_hook(self):
-        hooks = [
-            Hook({"path": "stacker.tests.test_util.mock_hook",
-                  "required": True, "enabled": False})]
-        handle_hooks("missing", hooks, self.provider, self.context)
-        self.assertTrue(hook_queue.empty())
-
-    def test_context_provided_to_hook(self):
-        hooks = [
-            Hook({"path": "stacker.tests.test_util.context_hook",
-                  "required": True})]
-        handle_hooks("missing", hooks, "us-east-1", self.context)
-
-    def test_hook_failure(self):
-        hooks = [
-            Hook({"path": "stacker.tests.test_util.fail_hook",
-                  "required": True})]
-        with self.assertRaises(SystemExit):
-            handle_hooks("fail", hooks, self.provider, self.context)
-        hooks = [{"path": "stacker.tests.test_util.exception_hook",
-                  "required": True}]
-        with self.assertRaises(Exception):
-            handle_hooks("fail", hooks, self.provider, self.context)
-        hooks = [
-            Hook({"path": "stacker.tests.test_util.exception_hook",
-                  "required": False})]
-        # Should pass
-        handle_hooks("ignore_exception", hooks, self.provider, self.context)
-
-    def test_return_data_hook(self):
-        hooks = [
-            Hook({
-                "path": "stacker.tests.test_util.result_hook",
-                "data_key": "my_hook_results"
-            }),
-            # Shouldn't return data
-            Hook({
-                "path": "stacker.tests.test_util.context_hook"
-            })
-        ]
-        handle_hooks("result", hooks, "us-east-1", self.context)
-
-        self.assertEqual(
-            self.context.hook_data["my_hook_results"]["foo"],
-            "bar"
-        )
-        # Verify only the first hook resulted in stored data
-        self.assertEqual(
-            list(self.context.hook_data.keys()), ["my_hook_results"]
-        )
-
-    def test_return_data_hook_duplicate_key(self):
-        hooks = [
-            Hook({
-                "path": "stacker.tests.test_util.result_hook",
-                "data_key": "my_hook_results"
-            }),
-            Hook({
-                "path": "stacker.tests.test_util.result_hook",
-                "data_key": "my_hook_results"
-            })
-        ]
-
-        with self.assertRaises(KeyError):
-            handle_hooks("result", hooks, "us-east-1", self.context)
 
 
 class TestException1(Exception):
